Вынесла окно пользователя в отдельный компонент
