+++ src/components/BaseButton.vue
@@ -1,5 +1,5 @@
 <template>
-  <button :class="['button', type ?? 'primary']" class="modal__btn-enter _hover01">
+  <button :class="['button', type ?? 'primary']">
     <RouterLink to="/">
       <slot></slot>
     </RouterLink>
@@ -12,32 +12,4 @@ defineProps({
 })
 </script>
 
-<style scoped>
-.modal__btn-enter {
-  width: 100%;
-  height: 30px;
-  background-color: #565eef;
-  border-radius: 4px;
-  margin-top: 20px;
-  margin-bottom: 20px;
-  border: none;
-  outline: none;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  font-size: 14px;
-  line-height: 21px;
-  font-weight: 500;
-  letter-spacing: -0.14px;
-  color: #ffffff;
-
-  & a {
-    width: 100%;
-    height: 100%;
-    color: #ffffff;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-  }
-}
-</style>
+<style scoped></style>
